investai-backend/requirements.txt
# ■■ Web Framework ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
# ■■ Database ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
# ■■ Validation & Settings ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
# ■■ Auth / Security ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.4.3
# ■■ HTTP / Scraping ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
# ■■ Celery / Redis ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
celery==5.4.0
redis==5.0.4
flower==2.0.1
firebase-admin==6.5.0
# ■■ Sentiment Analysis ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
vaderSentiment==3.3.2
transformers==4.41.1
# torch: uncomment for GPU/CPU inference (large download)
# torch==2.3.0 --index-url https://download.pytorch.org/whl/cpu
# ■■ AI Agent ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
anthropic==0.28.0 # NEW — Claude API SDK
google-generativeai==0.5.4
groq==0.9.0
openai==1.30.1
# ■■ Reliability & Logging ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
tenacity==8.3.0
structlog==24.1.0
# ■■ Dev / Testing ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0 # also used for test client
